requirements.txt
aiogTTS==1.1.1
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
arrow==1.2.2
async-google-trans-new==1.4.5
async-timeout==4.0.2
asyncgTTS==0.3.1
attrs==21.4.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2022.5.18.1
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
debtcollector==2.5.0
frozenlist==1.3.0
future==0.18.2
googlesearch-python==1.1.0
gTTS==2.2.4
humanize==4.1.0
idna==2.10
lxml==4.8.0
multidict==6.0.2
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
python-whois==0.7.3
pytz==2022.1
requests==2.25.1
six==1.16.0
soupsieve==2.3.2.post1
typing-extensions==4.2.0
urllib3==1.26.9
wavelink==1.2.5
wrapt==1.14.1
yarl==1.7.2
